@@ -1,24 +1,25 @@
 {
   "name": "yasmin_viewer_web_client",
-  "version": "0.1.0",
+  "version": "3.2.0",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.3.1",
-    "@mui/material": "^5.3.1",
+    "@mui/icons-material": "^6.4.0",
+    "@mui/material": "^6.4.0",
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^13.5.0",
-    "cytoscape": "^3.21.1",
-    "cytoscape-cola": "^2.5.1",
-    "cytoscape-dagre": "^2.4.0",
+    "cytoscape": "^3.31.0",
+    "cytoscape-cose-bilkent": "^4.1.0",
+    "cytoscape-dagre": "^2.5.0",
     "cytoscape-klay": "^3.1.4",
-    "react": "^17.0.2",
-    "react-cytoscapejs": "^1.2.1",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.4"
+    "react": "^19.0.0",
+    "react-cytoscapejs": "^2.0.0",
+    "react-dom": "^19.0.0",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^2.1.4",
+    "webpack": "^5.97.1"
   },
   "scripts": {
     "start": "react-scripts start",
