@@ -1,4 +1,4 @@
-# Copyright (C) 2023  Miguel Ángel González Santamarta
+# Copyright (C) 2023 Miguel Ángel González Santamarta
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -14,7 +14,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from typing import Any, Dict
-from threading import Lock
+from threading import RLock as Lock
 
 import yasmin
 
@@ -55,6 +55,8 @@ class Blackboard(object):
         self.__lock: Lock = Lock()
         ## A dictionary holding the data stored in the blackboard.
         self._data: Dict[str, Any] = {}
+        ## A dictionary containing all remappings
+        self.__remapping: Dict[str, str] = {}
 
         if init is not None:
             self._data.update(init)  # Initialize with provided data
@@ -75,7 +77,10 @@ class Blackboard(object):
         yasmin.YASMIN_LOG_DEBUG(f"Getting '{key}' from the blackboard")
 
         with self.__lock:
-            return self._data[key]
+            if not self.__contains__(key):
+                raise KeyError(f"Element '{key}' does not exist in the blackboard")
+
+            return self._data[self.__remap(key)]
 
     def __setitem__(self, key: str, value: Any) -> None:
         """
@@ -112,7 +117,7 @@ class Blackboard(object):
         yasmin.YASMIN_LOG_DEBUG(f"Removing '{key}' from the blackboard")
 
         with self.__lock:
-            del self._data[key]
+            del self._data[self.__remap(key)]
 
     def __contains__(self, key: str) -> bool:
         """
@@ -130,7 +135,7 @@ class Blackboard(object):
         yasmin.YASMIN_LOG_DEBUG(f"Checking if '{key}' is in the blackboard")
 
         with self.__lock:
-            return key in self._data
+            return self.__remap(key) in self._data
 
     def __len__(self) -> int:
         """
@@ -157,3 +162,45 @@ class Blackboard(object):
         """
         with self.__lock:
             return repr(self._data)
+
+    @property
+    def remappings(self) -> Dict[str, str]:
+        """
+        Property getter of a dict of remappings of the blackboard keys.
+
+        Return:
+            Dict: A dict with the current key remappings of the blackboard.
+
+        Raises:
+            None
+        """
+        return self.__remapping
+
+    @remappings.setter
+    def remappings(self, remapping: Dict[str, str]) -> None:
+        """
+        Property setter of the remapping of the blackboard keys.
+
+        Args:
+            remapping (Dict): The new remap of the blackboard.
+
+        Returns:
+            None
+
+        Raises:
+            None
+        """
+        self.__remapping = remapping
+
+    def __remap(self, key: str) -> str:
+        """
+        internal method that acquires the maped key. In the case the key is not remaped, retruns the arg key.
+
+        Args:
+            key (str): The key to be remaped.
+        Returns:
+            str: The remaped key or if is not remaped, the own key.
+        Raises:
+            None
+        """
+        return self.remappings[key] if key in self.remappings.keys() else key
