@@ -1,4 +1,4 @@
-// Copyright (C) 2023  Miguel Ángel González Santamarta
+// Copyright (C) 2023 Miguel Ángel González Santamarta
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -141,6 +141,7 @@ int main(int argc, char *argv[]) {
           {"outcome2", "outcome4"}, // Transition to outcome4 on outcome2
           {yasmin_ros::basic_outcomes::TIMEOUT,
            "outcome4"}, // Timeout transition
+          {yasmin_ros::basic_outcomes::CANCEL, "outcome4"}, // Cancel transition
       });
 
   // Publisher for visualizing the state machine's status
