@@ -1,4 +1,4 @@
-# Copyright (C) 2023  Miguel Ángel González Santamarta
+# Copyright (C) 2023 Miguel Ángel González Santamarta
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -13,17 +13,19 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import time
 from typing import List, Set, Callable, Union, Type, Any
+from threading import Event
 
 from rclpy.node import Node
 from rclpy.subscription import Subscription
 from rclpy.qos import QoSProfile
+from rclpy.callback_groups import CallbackGroup
 
+import yasmin
 from yasmin import State
 from yasmin import Blackboard
 from yasmin_ros.yasmin_node import YasminNode
-from yasmin_ros.basic_outcomes import TIMEOUT
+from yasmin_ros.basic_outcomes import TIMEOUT, CANCEL
 
 
 class MonitorState(State):
@@ -38,7 +40,6 @@ class MonitorState(State):
         msg_list (List[Any]): A set of messages received from the topic.
         msg_queue (int): The maximum number of messages to retain.
         time_to_wait (float): Time to wait between checks for messages.
-        monitoring (bool): Flag indicating if monitoring is active.
         _timeout (int): Timeout duration for monitoring.
 
     Exceptions:
@@ -52,6 +53,7 @@ class MonitorState(State):
         outcomes: Set[str],
         monitor_handler: Callable,
         qos: Union[QoSProfile, int] = 10,
+        callback_group: CallbackGroup = None,
         msg_queue: int = 10,
         node: Node = None,
         timeout: int = None,
@@ -65,6 +67,7 @@ class MonitorState(State):
             outcomes (Set[str]): The set of possible outcomes from the state.
             monitor_handler (Callable[[Blackboard, Any], None]): The function to call with the received messages.
             qos (Union[QoSProfile, int], optional): Quality of Service profile or depth.
+            callback_group (CallbackGroup, optional): The callback group for the subscription.
             msg_queue (int, optional): Maximum number of messages to store. Default is 10.
             node (Node, optional): The ROS node to use. If None, a default node is created.
             timeout (int, optional): Timeout in seconds for monitoring before giving up.
@@ -80,15 +83,14 @@ class MonitorState(State):
         ## The maximum number of messages to retain.
         self.msg_queue: int = msg_queue
         ## Time to wait between checks for messages.
-        self.time_to_wait: float = 0.001
-        ## Flag indicating if monitoring is active.
-        self.monitoring: bool = False
+        self._msg_event: Event = Event()
 
         ## Timeout duration for monitoring.
         self._timeout: int = timeout
 
         if timeout is not None:
             outcomes = [TIMEOUT] + outcomes
+        outcomes = [CANCEL] + outcomes
 
         ## The ROS 2 node instance used for subscriptions.
         self._node: Node = node
@@ -100,13 +102,23 @@ class MonitorState(State):
         self._topic_name: str = topic_name
 
         ## Subscription object for the specified topic.
+        self._sub: Subscription = None
+        self._msg_type: Type = msg_type
+        self._qos: Union[QoSProfile, int] = qos
+        self._callback_group: CallbackGroup = callback_group
+
+        ## Subscription object for the specified topic.
         self._sub: Subscription = self._node.create_subscription(
-            msg_type, topic_name, self.__callback, qos
+            self._msg_type,
+            self._topic_name,
+            self.__callback,
+            self._qos,
+            callback_group=self._callback_group,
         )
 
         super().__init__(outcomes)
 
-    def __callback(self, msg) -> None:
+    def __callback(self, msg: Any) -> None:
         """
         Callback function that handles incoming messages.
 
@@ -118,11 +130,12 @@ class MonitorState(State):
         Returns:
             None
         """
-        if self.monitoring:
-            self.msg_list.append(msg)
+        self.msg_list.append(msg)
 
-            if len(self.msg_list) >= self.msg_queue:
-                self.msg_list.pop(0)
+        if len(self.msg_list) > self.msg_queue:
+            self.msg_list.pop(0)
+
+        self._msg_event.set()
 
     def execute(self, blackboard: Blackboard) -> str:
         """
@@ -142,25 +155,32 @@ class MonitorState(State):
             None raised directly; timeouts are handled gracefully.
         """
 
-        elapsed_time: float = 0.0
-        self.msg_list: List[Any] = []
-        self.monitoring: bool = True
-
         # Wait until a message is received or timeout is reached
+        self._msg_event.clear()
+
         while not self.msg_list:
-            time.sleep(self.time_to_wait)
+            flag = self._msg_event.wait(self._timeout)
 
-            if self._timeout is not None:
-                if elapsed_time >= self._timeout:
-                    self.monitoring: bool = False
-                    self._node.get_logger().warn(
-                        f"Timeout reached, topic '{self._topic_name}' is not available"
-                    )
-                    return TIMEOUT
-                elapsed_time += self.time_to_wait
+            if self.is_canceled():
+                return CANCEL
 
-        self._node.get_logger().info(f"Processing msg from topic '{self._topic_name}'")
+            if self._timeout is not None and not flag:
+                yasmin.YASMIN_LOG_WARN(
+                    f"Timeout reached, topic '{self._topic_name}' is not available"
+                )
+                return TIMEOUT
+
+        yasmin.YASMIN_LOG_INFO(f"Processing msg from topic '{self._topic_name}'")
         outcome = self._monitor_handler(blackboard, self.msg_list.pop(0))
 
-        self.monitoring: bool = False
         return outcome
+
+    def cancel_state(self) -> None:
+        """
+        Cancels the current monitor state.
+
+        This method cancels the monitor if waiting for messages.
+        """
+
+        super().cancel_state()
+        self._msg_event.set()
